+++ src/Compute/Compute/ChangeLog.md
@@ -20,6 +20,7 @@
 
 -->
 ## Upcoming Release
+* Deprecated `Get-AzVMSize` 'List Virtual Machine Size' parameter set. 
 * Added new parameters `EnableAutomaticZoneRebalance`, `AutomaticZoneRebalanceStrategy` and `AutomaticZoneRebalanceBehavior` to `New-AzVmssConfig` and `Update-AzVmss` cmdlets for VMSS Automatic Zone Rebalancing. 
 
 ## Version 9.3.0